% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_binary_categorical_X}
\alias{assert_binary_categorical_X}
\title{Assert that X is numeric or (if categorical) has at most two levels}
\usage{
assert_binary_categorical_X(mod, X)
}
\arguments{
\item{mod}{An "rq"/"rqs" object}

\item{X}{Character scalar: name of the covariate of interest}
}
\description{
Works with rq/rqs even when model.frame or mod$x are not stored.
}
\keyword{internal}
