% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_intercept_present}
\alias{assert_intercept_present}
\title{Assert that the fitted model includes an intercept}
\usage{
assert_intercept_present(mod)
}
\arguments{
\item{mod}{An object of class "rq" or "rqs".}
}
\description{
Assert that the fitted model includes an intercept
}
\keyword{internal}
