% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_B_from_dist}
\alias{build_B_from_dist}
\title{Compute weights from error densities}
\usage{
build_B_from_dist(taus, error.distr, error.par = list())
}
\arguments{
\item{taus}{Numeric vector of quantile levels in (0,1).}

\item{error.distr}{Character. One of "normal", "skew-normal", "t".}

\item{error.par}{List of parameters for the chosen distribution.}
}
\description{
Compute weights from error densities
}
\keyword{internal}
