% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_dimensions_A}
\alias{check_dimensions_A}
\title{Assert that the matrix A used in the Imhof approximation has
valid dimensions at intersection level}
\usage{
check_dimensions_A(A, k)
}
\arguments{
\item{A}{The matrix A}

\item{k}{number of tests}
}
\description{
Assert that the matrix A used in the Imhof approximation has
valid dimensions at intersection level
}
\keyword{internal}
