% Generated by roxygen2 (4.0.1): do not edit by hand
\name{comp.rho}
\alias{comp.rho}
\title{rho calculation}
\usage{
comp.rho(ref, alt, eps)
}
\arguments{
\item{ref}{integer. reference allele count}

\item{alt}{integer. alternate allele count}

\item{eps}{float. error estimate \in [0,1]}
}
\value{
\rho float. elstimate of alternate allele frequency
}
\description{
\code(comp.rho) given reference and alternate read counts, & an error estimate, return the
estimated alternate alle frequency
}

