% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.build_A_from_dist}
\alias{.build_A_from_dist}
\title{Compute the matrix A from  error densities}
\usage{
.build_A_from_dist(taus, error.distr, error.par)
}
\arguments{
\item{taus}{The level of the quantiles}

\item{error.distr}{The distribution of the error terms (normal, exponential and t)}

\item{error.par}{The related parameters of the specified distribution}
}
\description{
Compute the matrix A from  error densities
}
\keyword{internal}
