% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.pImhof}
\alias{.pImhof}
\title{The Imhof method for x=0 and central variables only.}
\usage{
.pImhof(lams, x, eps = c(1e-10, 1e-10))
}
\arguments{
\item{lams}{eigenvalues}

\item{x}{observed stat test}

\item{eps}{tolerance vector (one for integrate, one for quadgk)}
}
\description{
The Imhof method for x=0 and central variables only.
}
