% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fitAseNull}
\alias{fitAseNull}
\title{fit QuASAR for a single sample}
\usage{
fitAseNull(ref, alt, eps = 0.1, log.gmat, max.it = 100, tol = 1e-16,
  fixGprior = TRUE)
}
\arguments{
\item{ref}{reference read count}

\item{alt}{alternate read count}

\item{eps}{starting value}

\item{log.gmat}{log of genotype priors}

\item{max.it}{maximum number of iterations for algorithm}

\item{tol}{tolerance to assess convergance}

\item{fixGprior}{logical to fix genotype priors across all steps of the algorithm}
}
\value{
list of genotypes, log genotypes, error estimate, log-likelihood, and the sum of
log likelihoods
}
\description{
\code(fitAseNull) for a single sample conduct genotyping and estimate sample error
}

