% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fitAseNullMulti}
\alias{fitAseNullMulti}
\title{fit QuASAR for a multiple samples}
\usage{
fitAseNullMulti(ref, alt, eps = rep(0.1, ncol(ref)), log.gmat, max.it = 100,
  tol = 1e-16, fixGprior = TRUE, verbose = TRUE)
}
\arguments{
\item{ref}{matrix of reference read count}

\item{alt}{matrix alternate read count}

\item{eps}{vector of starting values}

\item{log.gmat}{log of genotype priors}

\item{max.it}{maximum number of iterations for algorithm}

\item{tol}{tolerance to assess convergance}

\item{fixGprior}{logical to fix genotype priors across all steps of the algorithm}

\item{verbose}{logical turn on reporting during algorithm iterations}
}
\value{
list of genotypes, log genotypes, vector oferror estimates, log-likelihood, and the sum of
log likelihoods
}
\description{
\code(fitAseNull) joint geontype across samples and estimate sample error
}

