% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gLogLikBetaBinomial}
\alias{gLogLikBetaBinomial}
\title{gradient of log-likelihood of beta-binomial model}
\usage{
gLogLikBetaBinomial(logit.p, D, R, A)
}
\arguments{
\item{logit.p}{logit(rho)}

\item{D}{dispersion estimate}

\item{R}{reference read count}

\item{A}{alternate read count}
}
\value{
gradient evaulated at logit(rho)
}
\description{
returns the gradient of the log likelihood of reference & alternate
read count data given logit(rho) and adispersion estimate.
}

