% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lrtEpsRhoBinom}
\alias{lrtEpsRhoBinom}
\title{critical value for likelihood ratio test with misspecififed genotypes}
\usage{
lrtEpsRhoBinom(ref, alt, eps, rho)
}
\arguments{
\item{ref}{reference read count}

\item{alt}{alternate read count}

\item{eps}{estimate of error}

\item{rho}{\rho estimate}
}
\value{
llrt critical value
}
\description{
\code(lrtEpsRhoBinom) returns the critical value of llrt comparing with
H0: the het is mis-specified or \rho=o.5 & H1: the het has imbalance with estimated \rho
}

