% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PrepForGenotyping}
\alias{PrepForGenotyping}
\title{PrepForGenotyping}
\usage{
PrepForGenotyping(ase.dat, min.coverage, dampen.priors = TRUE)
}
\arguments{
\item{ase.dat}{data from UnionExtractFields.R}

\item{min.coverage}{minimumm coverage across all combined sample for locus inclusion}

\item{dampen.priors}{prevents underflow}
}
\value{
blah blah blah
}
\description{
PrepForGenotyping
}

