% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{qsr_init}
\alias{qsr_init}
\title{Initialize a new QUASAR project}
\usage{
qsr_init(
  name,
  type = c("academic", "bi_dashboard", "spark_pipeline"),
  author = getOption("quasar.author", default = ""),
  path = ".",
  journal = c("r_journal", "jss", "apsr")
)
}
\arguments{
\item{name}{Character. Name of the project. Use snake_case.}

\item{type}{Character. Type of project. One of "academic", "bi_dashboard",
or "spark_pipeline". Defaults to "academic".}

\item{author}{Character. Author name. Defaults to global option if set.}

\item{path}{Character. Where to create the project. Defaults to current directory.}

\item{journal}{Character. Target journal for academic output formatting.
One of "r_journal", "jss", "apsr". Defaults to "r_journal".}
}
\value{
Invisible path to the created project.
}
\description{
Creates a standardized project structure following QUASAR conventions.
This is the entry point of the QUASAR framework.
}
\examples{
\dontrun{
qsr_init("electoral_analysis", type = "academic", author = "Nache")
}
}
