% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{qsr_reset}
\alias{qsr_reset}
\title{Reset the QUASAR context}
\usage{
qsr_reset()
}
\value{
Invisible.
}
\description{
Clears all configuration, data, models, and outputs from the session.
}
\examples{
qsr_reset()
}
