% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasar-methods.R
\name{quasar-methods}
\alias{quasar-methods}
\alias{print.quasar}
\alias{summary.quasar}
\title{Print and summary methods for quasar objects}
\usage{
\method{print}{quasar}(x, ...)

\method{summary}{quasar}(object, ..., alpha = 0.05)
}
\arguments{
\item{x, object}{An object of class \code{quasar}.}

\item{...}{Additional arguments passed to other methods.}

\item{alpha}{Significance level.}
}
\value{
The input object invisibly.
}
\description{
These methods provide basic information about objects of class
\code{quasar}, typically returned by the \code{\link{closedTesting}} function.
}
\author{
Anna Vesely
}
