% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_describe}
\alias{rust_describe}
\title{Fast summary statistics for numeric columns.}
\usage{
rust_describe(path)
}
\arguments{
\item{path}{Character. Path to CSV file.}
}
\description{
Fast summary statistics for numeric columns.
}
