% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_filter}
\alias{rust_filter}
\title{Fast filter rows.}
\usage{
rust_filter(path, col_name, op, value)
}
\arguments{
\item{path}{Character. Path to CSV file.}

\item{col_name}{Character. Column to filter on.}

\item{op}{Character. One of: gt, lt, gte, lte, eq, neq.}

\item{value}{Double. Value to compare against.}
}
\description{
Fast filter rows.
}
