% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_group_agg}
\alias{rust_group_agg}
\title{Fast group-by aggregation.}
\usage{
rust_group_agg(path, group_col, agg_col, agg_fn)
}
\arguments{
\item{path}{Character. Path to CSV file.}

\item{group_col}{Character. Column to group by.}

\item{agg_col}{Character. Column to aggregate.}

\item{agg_fn}{Character. One of: mean, sum, count, min, max, std.}
}
\description{
Fast group-by aggregation.
}
