% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_read_parquet}
\alias{rust_read_parquet}
\title{Read a Parquet file using Polars.}
\usage{
rust_read_parquet(path)
}
\arguments{
\item{path}{Character. Path to the Parquet file.}
}
\description{
Read a Parquet file using Polars.
}
