% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_sort}
\alias{rust_sort}
\title{Fast sort rows.}
\usage{
rust_sort(path, by, descending)
}
\arguments{
\item{path}{Character. Path to CSV file.}

\item{by}{Character. Column to sort by.}

\item{descending}{Logical.}
}
\description{
Fast sort rows.
}
