% Generated by roxygen2 (4.0.1): do not edit by hand
\name{UnionExtractFields}
\alias{UnionExtractFields}
\title{UnionExtractFields}
\usage{
UnionExtractFields(fileList, combine = FALSE)
}
\arguments{
\item{fileList}{list of clean bed files as described in the QuASAR vingette}
}
\value{
blah blah blah
}
\description{
from a list of clean bed files, find the intersection of all loci and return a list with reference, alternate, and error counts for each loci
across all samples
}

