% Generated by roxygen2 (4.0.1): do not edit by hand
\name{aseInference}
\alias{aseInference}
\title{aseInference}
\usage{
aseInference(gts, priors, ref.mat, alt.mat, eps.vect, min.cov, sample.names)
}
\arguments{
\item{gts}{posterior probabilities of genotypes from QuASAR}

\item{priors}{1K genomes minor allele frequencies as priors}

\item{ref.mat}{matrix of reference allele counts}

\item{alt.mat}{matrix of alternate allele counts}

\item{eps.vect}{QuASAR estimates of sequencing error for each sample}

\item{min.cov}{threshold for the minimum coverage across all samples}

\item{sample.names}{verctor of sample names}
}
\value{
inference.data list.
}
\description{
using genotypes from QuASAR, conduct inference on allelic imbalance
}

