% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closedTesting.R
\name{closedTesting}
\alias{closedTesting}
\title{Closed Testing procedure}
\usage{
closedTesting(res, quantiles)
}
\arguments{
\item{res}{data.frame object from \code{\link{waldTest}}, \code{\link{rankTest}}}

\item{quantiles}{numeric vector that contains the set of quantiles of interest. If \code{NULL}
all single hypotheses are considered.}
}
\value{
Data.frame containing the quantile set defined in \code{quantiles} and the related adjusted \eqn{p}-values.
}
\description{
This function applies the closed testing procedure to strong control
of the FWER to test the effect of a covariate of interest when multiple
quantile regressions are fitted.
}
\examples{
library(quantreg)
tau=c(0.1,0.25,0.5,0.75,0.9)
x<-rnorm(100)
z<-rnorm(100)
y<-rnorm(100)
mod <- rq(y ~ x + z, tau = tau)
res <- waldTest(mod = mod,  X = "x")
closedTesting(res, quantiles = c(0.1))
}
\references{
Marcus, R., Eric, P., & Gabriel, K. R. (1976). On closed testing procedures with special reference to ordered analysis of variance. Biometrika, 63(3), 655-660.

Goeman, J. J., Hemerik, J., & Solari, A. (2021). Only closed testing procedures are admissible for controlling false discovery proportions. The Annals of Statistics, 49(2), 1218-1238.
}
\seealso{
\code{\link{waldTest}}, \code{\link{rankTest}}
}
\author{
Angela Andreella
}
