% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.pImhof}
\alias{.pImhof}
\title{The Imhof method for x=0 and central variables only.}
\usage{
.pImhof(lams, eps = c(1e-04, 1e-04), x)
}
\arguments{
\item{lams}{eigenvalues}

\item{eps}{tolerance vector (one for integrate, one for quadgk)}

\item{x}{observed stat test}
}
\description{
The Imhof method for x=0 and central variables only.
}
