% Generated by roxygen2 (4.0.1): do not edit by hand
\name{logLikBetaBinomial2}
\alias{logLikBetaBinomial2}
\title{log-likelihood of beta-binomial model for heterozygotes}
\usage{
logLikBetaBinomial2(logit.p, D, R, A)
}
\arguments{
\item{rho}{rho value in [0, 1]}

\item{eps}{estimate of error}

\item{D}{dispersion estimate}

\item{R}{reference read count}

\item{A}{alternate read count}
}
\value{
log likelihood
}
\description{
returns the log likelihood of reference & alternate read count data
given rho, & dispersion.
}

