% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasar-methods.R
\name{quasar-methods}
\alias{quasar-methods}
\alias{summary.quasar}
\alias{print.quasar}
\title{Print and summary methods for quasar objects}
\usage{
\method{summary}{quasar}(x, alpha = 0.05, ...)

\method{print}{quasar}(x, ...)
}
\arguments{
\item{x}{An object of class \code{quasar}.}

\item{alpha}{Significance level.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
The input object invisibly.
}
\description{
These methods provide basic information about objects of class
\code{quasar}, typically returned by the \code{\link{closedTesting}} function.
}
\author{
Anna Vesely
}
