% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankTest.R
\name{rankTest}
\alias{rankTest}
\title{Rank score test for quantile regression}
\usage{
rankTest(mod, X, h = NULL)
}
\arguments{
\item{mod}{\code{rqs} object from \code{rq} function (\code{quantreg} package) that is the
quantile models of interest}

\item{X}{string value that is the covariate of interest}

\item{h}{bandwidth value, default \code{NULL}}
}
\value{
Data.frame containing the quantile set, statistical test and related \eqn{p}-value
}
\description{
This function computes the rank score test
}
\examples{
library(quantreg)
tau=c(0.1,0.25,0.5,0.75,0.9)
x<-rnorm(100)
z<-rnorm(100)
y<-rnorm(100)
mod <- rq(y ~ x + z, tau = tau)
rankTest(mod = mod,  X = "x")
}
\seealso{
\code{\link{rankTest}}
}
\author{
Angela Andreella
}
