% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_read_csv}
\alias{rust_read_csv}
\title{Read a CSV file using Polars (10-30x faster than read.csv).}
\usage{
rust_read_csv(path, n_rows)
}
\arguments{
\item{path}{Character. Path to the CSV file.}

\item{n_rows}{Integer or NULL. Max rows to read.}
}
\description{
Read a CSV file using Polars (10-30x faster than read.csv).
}
