% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateData}
\alias{simulateData}
\title{simulate data}
\usage{
simulateData(n, beta, gamma, mu, Sigma, sigma.y)
}
\arguments{
\item{n}{number of observations}

\item{beta}{value of the parameter of interest}

\item{gamma}{vector of values of the nuisance parameters}

\item{mu}{vector of means for covariates}

\item{Sigma}{covariance matrix for covariates}

\item{sigma.y}{vector of standard deviations for the response variable}
}
\value{
matrix
}
\description{
This function simulate data
}
\author{
Angela Andreella
}
