% Generated by roxygen2 (4.0.1): do not edit by hand
\name{UnionExtractFields}
\alias{UnionExtractFields}
\title{UnionExtractFields}
\usage{
UnionExtractFields(fileList, combine = FALSE)
}
\arguments{
\item{fileList}{List of files *.quasar.in.gz}

\item{combine}{Collapses all samples into a single sample if true. Default is false}
}
\value{
returns a R list with the following elements:
Ref: Matrix with number of reads matching the reference allele.
Alt: Matrix with number of reads matching the alternate allele.
Err: Matrix with number of reads matchine neither the ref. or alt. allele.
anno: Object with the annotation for the SNPs.
}
\description{
From a list of files (fileList), find the union of all loci and return a list with the data prepared to run QuASAR.
}
\author{
Chris Harvey
}

