% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waldTest.R
\name{waldTest}
\alias{waldTest}
\title{Wald-type test for quantile regression}
\usage{
waldTest(mod, X, tau = NULL, full = FALSE, h = NULL, beta = 0, alpha = 0.05)
}
\arguments{
\item{mod}{An object of class \code{"rqs"} returned by
\code{\link[quantreg]{rq}}, representing the fitted quantile regression models.}

\item{X}{A string indicating the covariate of interest.}

\item{tau}{A numeric vector of quantiles of interest used in \code{mod}.
Default \code{NULL}, i.e., it takes all the quantiles from the \code{mod} object.}

\item{full}{Logical. If \code{TRUE}, the function returns the test statistics
and corresponding \eqn{p}-values for all intersection hypotheses containing
\code{tau}. If \code{FALSE} (default), only the results for the single hypotheses
are returned.}

\item{h}{A numeric value for the bandwidth. Default is \code{NULL}.}

\item{beta}{numeric value that is the value of the parameter of interest under the null hypothesis
Default 0.}

\item{alpha}{A numeric value used for bandwidth estimation.
Following Koenker (2005), it is typically set equal to the desired \eqn{\alpha} level.
The default is 0.05.}
}
\value{
A \code{data.frame} containing:
\itemize{
  \item \code{Quantiles.Set}: the quantiles set
  \item \code{Statistic}: the Wald-type test statistic
  \item \code{p.value}: the corresponding unadjusted \eqn{p}-value
}
}
\description{
Performs the Wald-type test for the covariate of interest specified in
\code{X}, at the quantiles defined in \code{tau}, using a fitted quantile
regression model.
}
\details{
This procedure requires that the covariate of interest \code{X} is either numeric
or, if categorical, has at most two levels. Multilevel categorical covariates
are not supported and will trigger an error.
}
\examples{
library(quantreg)
tau=c(0.1,0.25,0.5,0.75,0.9)
x<-rnorm(100)
z<-rnorm(100)
y<-rnorm(100)
mod <- rq(y ~ x + z, tau = tau)
waldTest(mod = mod,  X = "x")
}
\seealso{
\code{\link[quantreg]{rq}} \code{\link[quasar]{rankTest}}
}
\author{
Angela Andreella
}
